fixed currency issue

+++ upload/admin/model/localisation/currency.php
@@ -139,7 +139,7 @@ class ModelLocalisationCurrency extends Model {
 			
 			if ((float)$value < 1 && isset($line[$i + 1])) {
 				if((float)utf8_substr($line[$i + 1], 11, 6) > 0) {
-					$value = (1 / utf8_substr($line[$i + 1], 11, 6));
+					$value = (1 / (float)utf8_substr($line[$i + 1], 11, 6));
 				} else {
 					$value = 0;
 				}
